backprop kinda working
