+++ README.md
@@ -2,11 +2,13 @@
 
 This is a python library made to help you build machine learning models.
 
-## Quicksttart
+## Quickstart
+lll
+
 
 ```python
 from pynn.layers import Dense, Sigmoid
-from pynn.ann import NeuralNetwork
+from pynn.network import NeuralNetwork
     
 n_weights_1 = 5 # 5 neurons in the first layer
 n_weights_2 = 1 # 1 neuron in the second layer (output)
@@ -22,10 +24,9 @@ nn = NeuralNetwork([
 
 - [X] Classes definition and construction
 - [X] Forward propagation fully working
-- [ ] Backward propagation working
-- [ ] Train and predict fully working
+- [X] Backward propagation working
+- [X] Train and predict fully working
 - [ ] Add Optimization layers
 - [ ] Add convolutional layers
 - [ ] Add batch support for SGD
 - [ ] Add native pyplot support
-- [ ] Add CUDA support??
